chore(required-attrs): make work with standards object
