+++ lib/checks/aria/aria-required-attr-evaluate.js
@@ -1,48 +1,18 @@
-import {
-	isNativeTextbox,
-	isNativeSelect,
-	isAriaTextbox,
-	isAriaListbox,
-	isAriaCombobox,
-	isAriaRange
-} from '../../commons/forms';
 import { requiredAttr } from '../../commons/aria';
+import { getElementSpec } from '../../commons/standards';
 import { uniqueArray, getNodeFromTree } from '../../core/utils';
-import matches from '../../commons/matches';
 
 function ariaRequiredAttrEvaluate(node, options = {}) {
 	const vNode = getNodeFromTree(node);
 	const missing = [];
 
-	const preChecks = {
-		// aria-valuenow should fail if element does not have a value property
-		// @see https://github.com/dequelabs/axe-core/issues/1501
-		'aria-valuenow': function() {
-			return !(
-				isNativeTextbox(node) ||
-				isNativeSelect(node) ||
-				isAriaTextbox(node) ||
-				isAriaListbox(node) ||
-				isAriaCombobox(node) ||
-				(isAriaRange(node) && node.hasAttribute('aria-valuenow'))
-			);
-		},
-		// aria-checked should fail if element does not have a checked property
-		// @see https://github.com/dequelabs/axe-core/issues/2225
-		'aria-checked': function() {
-			return !matches(vNode, {
-				nodeName: 'input',
-				attributes: {
-					type: ['checkbox', 'radio']
-				}
-			});
-		}
-	};
-
 	if (node.hasAttributes()) {
 		const role = node.getAttribute('role');
 		let required = requiredAttr(role);
+		const elmSpec = getElementSpec(vNode);
 
+		// @deprecated: required attr options to pass more attrs.
+		// configure the standards spec instead
 		if (Array.isArray(options[role])) {
 			required = uniqueArray(options[role], required);
 		}
@@ -51,7 +21,7 @@ function ariaRequiredAttrEvaluate(node, options = {}) {
 				const attr = required[i];
 				if (
 					!node.getAttribute(attr) &&
-					(preChecks[attr] ? preChecks[attr]() : true)
+					!(elmSpec.implicitAttrs && elmSpec.implicitAttrs[attr])
 				) {
 					missing.push(attr);
 				}
