+++ lib/commons/aria/required-attr.js
@@ -1,4 +1,4 @@
-import lookupTable from './lookup-table';
+import standards from '../../standards';
 
 /**
  * Get required attributes for a given role
@@ -9,9 +9,13 @@ import lookupTable from './lookup-table';
  * @return {Array}
  */
 function requiredAttr(role) {
-	const roles = lookupTable.role[role];
-	const attr = roles && roles.attributes && roles.attributes.required;
-	return attr || [];
+	const roleDef = standards.ariaRoles[role];
+
+	if (!roleDef || !roleDef.requiredAttrs) {
+		return [];
+	}
+
+	return roleDef.requiredAttrs;
 }
 
 export default requiredAttr;
