+++ test/integration/rules/aria-required-attr/required-attr.html
@@ -34,9 +34,10 @@
 <span role="radio" id="pass8" aria-checked="true">I am GREEN!</span>
 <input type="number" role="spinbutton" id="pass9" value="10" />
 <input type="number" role="spinbutton" id="pass10" />
+<div role="scrollbar" id="pass11">ok</div>
+<div role="slider" id="pass12">ok</div>
+<div role="spinbutton" id="pass13">ok</div>
 
-<div role="scrollbar" id="violation1">fail</div>
-<div role="slider" id="violation2">fail</div>
-<div role="spinbutton" id="violation3">fail</div>
-<div role="heading" id="violation4">fail</div>
-<div role="combobox" id="violation5">fail</div>
+<div role="heading" id="violation1">fail</div>
+<div role="combobox" id="violation2">fail</div>
+<div role="switch" id="violation3">fail</div>
